dev_requirements.txt
strawberry-graphql
strawberry-graphql-django
django
pytest
pytest-django
pytest-cov
